Does the successful methods
Refactored to represent this method relates to topics instead of projects.
Will now ask the NetworkConnManager for a list of topics.
onSuccess will format this correctly and forward it to the correct fragment.

Might want this class to directly implement Callback (which really should have a better name!) instead of relegating that to a subclass.